Some cleanups
Removed some unused and moved some locally used constants away from constants.c .
Removed CompactTuple methods from helpers
Changed arraylist resizing macro to a function

+++ tf2_dem_py/demo_parser/parser_state/data_structs/game_events.h
@@ -19,8 +19,8 @@ typedef struct {
 typedef struct {
 	uint16_t event_type;
 	uint8_t *name;
-	uint16_t entries_capacity;
-	uint16_t entries_len;
+	size_t entries_capacity;
+	size_t entries_len;
 	GameEventEntry *entries;
 } GameEventDefinition;
